Refactored to use poetry instead of setup.py
- Preserved entrypoint run.py for direct/unpackaged use
- Moved bids.py to utils folder to avoid circular import
- Moved functionality from run.py to extract_tacs.py
- Created cli() method in extract_tacs.py to provide entry point
for poetry/package installation in lieu of __name__ == "main"
- bumped version to 0.0.4 to account for minor changes
